Separate Code and Markdown View Model
