+++ src/vs/workbench/contrib/notebook/browser/viewModel/notebookViewModel.ts
@@ -3,22 +3,23 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import { onUnexpectedError } from 'vs/base/common/errors';
 import { Emitter, Event } from 'vs/base/common/event';
 import { Disposable, DisposableStore } from 'vs/base/common/lifecycle';
-import * as editorCommon from 'vs/editor/common/editorCommon';
-import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
-import { NotebookEditorModel } from 'vs/workbench/contrib/notebook/browser/notebookEditorInput';
-import { CellViewModel } from 'vs/workbench/contrib/notebook/browser/viewModel/notebookCellViewModel';
-import { ICell } from 'vs/workbench/contrib/notebook/common/notebookCommon';
-import { IModelDeltaDecoration } from 'vs/editor/common/model';
-import { onUnexpectedError } from 'vs/base/common/errors';
-import { CellFindMatch, CellState, ICellViewModel } from 'vs/workbench/contrib/notebook/browser/notebookBrowser';
+import { URI } from 'vs/base/common/uri';
 import { IBulkEditService } from 'vs/editor/browser/services/bulkEditService';
 import { Range } from 'vs/editor/common/core/range';
+import * as editorCommon from 'vs/editor/common/editorCommon';
+import { IModelDeltaDecoration } from 'vs/editor/common/model';
 import { WorkspaceTextEdit } from 'vs/editor/common/modes';
-import { URI } from 'vs/base/common/uri';
+import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
 import { IUndoRedoService } from 'vs/platform/undoRedo/common/undoRedo';
-import { InsertCellEdit, DeleteCellEdit, MoveCellEdit } from 'vs/workbench/contrib/notebook/browser/viewModel/cellEdit';
+import { CellFindMatch, CellState, ICellViewModel } from 'vs/workbench/contrib/notebook/browser/notebookBrowser';
+import { NotebookEditorModel } from 'vs/workbench/contrib/notebook/browser/notebookEditorInput';
+import { DeleteCellEdit, InsertCellEdit, MoveCellEdit } from 'vs/workbench/contrib/notebook/browser/viewModel/cellEdit';
+import { CodeCellViewModel } from 'vs/workbench/contrib/notebook/browser/viewModel/codeCellViewModel';
+import { MarkdownCellViewModel } from 'vs/workbench/contrib/notebook/browser/viewModel/markdownCellViewModel';
+import { CellKind, ICell } from 'vs/workbench/contrib/notebook/common/notebookCommon';
 
 export interface INotebookEditorViewState {
 	editingCells: { [key: number]: boolean };
@@ -110,15 +111,15 @@ export class NotebookViewModel extends Disposable {
 			this._onDidChangeViewCells.fire({
 				synchronous: true,
 				splices: e.map(splice => {
-					return [splice[0], splice[1], splice[2].map(cell => this.instantiationService.createInstance(CellViewModel, this.viewType, this.handle, cell))];
+					return [splice[0], splice[1], splice[2].map(cell => {
+						return createCellViewModel(this.instantiationService, this.viewType, this.handle, cell);
+					})];
 				})
 			});
 		}));
 
 		this._viewCells = this._model!.notebook!.cells.map(cell => {
-			const viewCell = this.instantiationService.createInstance(CellViewModel, this.viewType, this._model!.notebook!.handle, cell);
-			this._localStore.add(viewCell);
-			return viewCell;
+			return createCellViewModel(this.instantiationService, this.viewType, this._model!.notebook!.handle, cell);
 		});
 	}
 
@@ -152,7 +153,7 @@ export class NotebookViewModel extends Disposable {
 	}
 
 	insertCell(index: number, cell: ICell, synchronous: boolean): CellViewModel {
-		const newCell = this.instantiationService.createInstance(CellViewModel, this.viewType, this.handle, cell);
+		let newCell: CellViewModel = createCellViewModel(this.instantiationService, this.viewType, this.handle, cell);
 		this._viewCells!.splice(index, 0, newCell);
 		this._model.insertCell(newCell.cell, index);
 		this._localStore.add(newCell);
@@ -380,5 +381,15 @@ export class NotebookViewModel extends Disposable {
 		});
 
 		super.dispose();
+	}
+}
+
+export type CellViewModel = CodeCellViewModel | MarkdownCellViewModel;
+
+export function createCellViewModel(instantiationService: IInstantiationService, viewType: string, notebookHandle: number, cell: ICell) {
+	if (cell.cellKind === CellKind.Code) {
+		return instantiationService.createInstance(CodeCellViewModel, viewType, notebookHandle, cell);
+	} else {
+		return instantiationService.createInstance(MarkdownCellViewModel, viewType, notebookHandle, cell);
 	}
 }
